Added sGroupMove and automatic handling of clash names while moving a group among zones.
What is missing now is that the HZone should inject code in sGroupMove to check that, when called on a Zone (when @GroupId is actually a Zone), we already have @GroupId.HierarchicalId.GetAncestor(1) = @NewZoneId.HierarchicalId.
If not, this means that we are not coming from sZoneMove and we are f$£* the parent/child relationships.

Making this transparent to the user implies another call level:
- When sGroupMove detects that the group is a zone:
   - if @GroupId.HierarchicalId.GetAncestor(1) = @NewZoneId.HierarchicalId, it continues its job
   - else it calls sZoneMove (with no @NextSiblingId, the zone will be appended)
      - sZoneMove will call it back with the HierarchicalId set.

This is the only way without "internal" procedures that we want to avoid as much as possible.

This mechanism is between HZone and Zone. It has nothing to do with names handling. We still don't need "optional dependency management" here as I initially thought...